Corrected exit status codes and improved comments in install and uninstall.sh, in uninstall.sh: updated warning_art for readability, install.sh added more error handling

+++ install.sh
@@ -12,27 +12,35 @@ who_run() {
     else
         echo "X You are NOT root!"
         echo "- Please run with root since some parts of this script interact with files in '/usr/local/bin/'."
-        exit 0
+        exit 1
     fi
 }
 
 who_run
 
-#Check if qwer is already installed by using command -v to search directories listed in $PATH including /usr/local/bin to see if there is an executable file called qwer
+#Check if qwer is already installed to and then if yes exits
 
 if ! command -v qwer >/dev/null 2>&1; then
     echo "- Installing now"
 else 
     echo "X qwer is already installed"
-    exit 0
+    exit 1
 fi
 
+#Check if qwer.sh exists in dir or else gives error
 
-#Copying qwer.sh as qwer to /usr/local/bin to make it executable 
-cp qwer.sh /usr/local/bin/qwer 
+if [ -f "./qwer.sh" ]; then
+    # Copying qwer.sh as qwer to /usr/local/bin to make it executable 
+    cp qwer.sh /usr/local/bin/qwer 
 
-chmod +x /usr/local/bin/qwer
+    chmod +x /usr/local/bin/qwer
 
-echo "- qwer has now been installed!"
-echo "- It can now be used by anyone, anywhere, just by typing 'qwer'."
-echo "- For more info, run 'qwer --help'."
+    echo "- qwer has now been installed!"
+    echo "- It can now be used by anyone, anywhere, just by typing 'qwer'."
+    echo "- For more info, run 'qwer --help'."
+else
+    echo "X Could not install since qwer.sh is not here."
+    echo "If you renamed it to something else, change it back to 'qwer.sh' and try again."
+    echo "Or, if you deleted it, please delete this whole directory and reclone the repo."
+    exit 1
+fi
